launch in bin64 when using account info
When launching with account info, we use nopatchui in order to launch the game directly without showing the window. Unfortunately, nopatchui causes the game to not change directory into bin64 before fully launching. Fix this by launching the game directly in bin64 using the working directory option.

It might be worth actually testing whether or not our arguments have nopatchui instead of relying on account name, but I haven't found a good way to do that yet.