Single launch account .dat file switching mockup
A mockup of .dat file symlinking for single accounts. Some things are disabled or bugged, but I figured you wouldn't mind having a somewhat functioning version for launching single accounts. No idea how you are going to be able to do this with multiple accounts. How does sandboxie work with GW2?

Changes...
The process priority setter checks if null, but not if normal (default). No point in changing normal to normal. This probably is still the wrong code, I haven't done any programming since my last commit.

Added checkbox and setting for datlaunching, this should probably be moved to the settings page, not the main launcher tab.

Added double click to launch an account. Only works like that when datlaunchingcheckbox has been checked.

Removed the line notifying user that their crash log doesn't exist.

Messing with the mutex is useless if launching only a single account. It now checks for that.

ToDo...
Need to add some sort of error check to see if a .dat file is actually linked when launching, this is buggy without.

Stop the launcher asking for a relaunch when launching with the -image parameter.