fix localdat gw2build id for DEBUG builds
If Gw2 Launchbuddy is built using the DEBUG configuration, it will not
properly update the local.dat Gw2 build ids. This unnecessarily causes
the program to continuously try to update the login files every time it
loads an account.

Fix this by removing the !DEBUG check around the id update code.

It does not make sense for the program to behave significantly
differently when in DEBUG mode.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>